+++ cdk/kittyhawk/src/application/base.ts
@@ -36,7 +36,8 @@ export class Application extends Construct {
     // We want to prepend the project name to the name of each component
     const release_name = process.env.RELEASE_NAME || "undefined_release";
     const fullname = `${release_name}-${appname}`;
-
+    const deploy_to_feature_branch =
+      process.env.DEPLOY_TO_FEATURE_BRANCH == "true";
     new Service(this, fullname, props.port);
 
     if (props.createServiceAccount) {
@@ -48,6 +49,14 @@ export class Application extends Construct {
     new Deployment(this, fullname, {
       ...props.deployment,
       port: props.port,
+      ...(deploy_to_feature_branch
+        ? {
+            env: [
+              ...(props.deployment.env ?? []),
+              { name: "DEPLOY_TO_FEATURE_BRANCH", value: "true" },
+            ],
+          }
+        : {}),
       ...(props.createServiceAccount
         ? { serviceAccountName: release_name }
         : {}),